requirements.txt
# Core dependencies
python-dotenv==1.0.0
pyyaml==6.0.1

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# PDF extraction
# camelot-py[cv]==0.11.0  # Incompatible with Python 3.12
PyMuPDF==1.23.21
pdfplumber==0.10.3
tabula-py==2.9.0

# Data processing
pandas==2.2.0
numpy==1.26.3
openpyxl==3.1.2

# Chemical informatics (optional)
# rdkit==2023.9.4  # Uncomment if structure-based matching needed

# Text matching
python-Levenshtein==0.23.0
fuzzywuzzy==0.18.0

# Sentence transformers for semantic matching
sentence-transformers>=2.3.1
torch>=2.2.0
transformers>=4.37.2

# FAISS for fast similarity search
faiss-cpu>=1.7.4  # Use faiss-gpu if CUDA available

# API clients
requests==2.31.0
requests-cache==1.1.1
ratelimit==2.2.1

# Testing
pytest==8.0.0
pytest-cov==4.1.0
pytest-mock==3.12.0

# Code quality
black==24.1.1
ruff==0.2.0
mypy==1.8.0

# Progress bars
tqdm==4.66.1

# Logging
loguru==0.7.2

# Scientific computing
scikit-learn==1.4.0
scipy==1.12.0
